Added support for uploading videos
Removed react-image and used a custom made component instead.
Switched to mime module to get file type from path instead of using file-type to get it from the buffer.
Added a link dropdown menu which copies a invitation link to the clipboard of the user.
Scroll to the bottom will be checked, if it's needed to apper on every text sent, received and deleted.