Added E2EE
When a new user joins a room, if that room doesn't exist it gets created with the password they provided and everyone else that joins that room will have to give the same password.
When the room is created they first user's client hashes it and then sends it and on the server it is hashed again and is saved. When another user joins the room the same thing happens but this time the 'double hash' is checked with one stored with the room.